02-HF/requirements.txt
# This file contains the required Python packages for the Hugging Face labs.
# To install them, run the following command in your terminal:
# pip install -r requirements.txt

transformers
datasets
torch
accelerate
scikit-learn
Pillow